My betty

+++ 0x01-variables_if_else_while/102-print_comb5.c
@@ -8,7 +8,7 @@ int main(void)
 {
 	int digi1, digi2;
 	int dcomaspace[] = {44, 32};
-	int z;
+	unsigned long int z;
 
 	for (digi1 = 0; digi1 <= 99; digi1++)
 	{
